+++ src/pages/Themes/components/ProductList.tsx
@@ -1,80 +1,18 @@
 import Loading from "@/components/common/Loading";
 import { ROUTE_PATH } from "@/components/routes/routePath";
-import useFetch from "@/hooks/useFetch";
+import usePaginationFetch from "@/hooks/usePaginationFetch";
 import type { RankingProductType } from "@/types/RankingProductType";
 import styled from "@emotion/styled";
-import { useCallback, useEffect, useRef, useState } from "react";
 import { Link, useParams } from "react-router-dom";
-
-interface ProductListData {
-  list: RankingProductType[];
-  cursor: number;
-  hasMoreList: boolean;
-}
 
 const PRODUCT_LIST_LIMIT = 20;
 
 const ProductList = () => {
-  const [items, setItems] = useState<RankingProductType[]>([]);
-  const [cursor, setCursor] = useState(0);
-  const [hasMoreList, setHasMoreList] = useState(true);
-  const [isLoading, setIsLoading] = useState(false);
-  const loader = useRef<HTMLDivElement>(null);
-
   const { themeId } = useParams();
-  const { fetchData } = useFetch<ProductListData>(`/api/themes/${themeId}/products`, {
-    autoFetch: false,
-  });
-
-  const loadMoreItems = useCallback(async () => {
-    if (isLoading || !hasMoreList) return null;
-
-    try {
-      setIsLoading(true);
-      const response = await fetchData(undefined, undefined, {
-        cursor,
-        limit: PRODUCT_LIST_LIMIT,
-      });
-
-      if (response.error) {
-        setHasMoreList(false);
-        return null;
-      }
-
-      if (response.data) {
-        setItems((prev) => [...prev, ...(response.data?.list ?? [])]);
-        setCursor(response.data.cursor);
-        setHasMoreList(response.data.hasMoreList);
-      }
-    } catch (error) {
-      console.error("상품 목록을 불러오는데 실패했습니다:", error);
-      setHasMoreList(false);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [cursor, hasMoreList, isLoading, fetchData]);
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      ([entries]) => {
-        if (entries.isIntersecting && hasMoreList && !isLoading) {
-          loadMoreItems();
-        }
-      },
-      { threshold: 0.5 },
-    );
-
-    const el = loader.current;
-    if (el && hasMoreList) {
-      observer.observe(el);
-    }
-
-    return () => {
-      if (el) {
-        observer.unobserve(el);
-      }
-    };
-  }, [cursor, hasMoreList, loadMoreItems, isLoading]);
+  const { items, isLoading, hasMoreList, loader } = usePaginationFetch<RankingProductType>(
+    `/api/themes/${themeId}/products`,
+    PRODUCT_LIST_LIMIT,
+  );
 
   return (
     <Container>
@@ -163,7 +101,6 @@ const Empty = styled.div`
   align-items: center;
   color: ${({ theme }) => theme.color.textColor.sub};
 `;
-
 const EndMessage = styled.div`
   width: 100%;
   font: ${({ theme }) => theme.typography.subtitle2Regular};
